reorg /moderate and /admin
